--- src/main/java/org/noostak/auth/application/KakaoService.java
@@ -1,5 +1,0 @@
-package org.noostak.auth.application;
-
-public interface KakaoService extends OauthService {
-
-}
