+++ src/main/java/org/noostak/auth/application/kakao/KaKaoApi.java
@@ -1,4 +1,4 @@
-package org.noostak.auth.application;
+package org.noostak.auth.application.kakao;
 
 
 import lombok.AllArgsConstructor;
