+++ src/main/java/org/noostak/auth/application/kakao/KakaoService.java
@@ -1,0 +1,7 @@
+package org.noostak.auth.application.kakao;
+
+import org.noostak.auth.application.OauthService;
+
+public interface KakaoService extends OauthService {
+
+}
